requirements.txt
pandas==2.0.3
requests==2.31.0
SQLAlchemy==2.0.20
psycopg2==2.9.7
tqdm==4.66.1
openpyxl==3.1.2
python-dotenv==1.0.0